- Moves AdminLogPlugin into separate file - Adds admin/user authorization level consts
